requirements.txt
torch>=2.8.0
numpy
Pillow
requests